Move createTimestampFile out of helpers
