+++ extensions/ql-vscode/src/local-queries/local-queries.ts
@@ -16,7 +16,6 @@ import { extLogger, TeeLogger } from "../common";
 import { isCanary, MAX_QUERIES } from "../config";
 import { gatherQlFiles } from "../pure/files";
 import { basename } from "path";
-import { createTimestampFile } from "../helpers";
 import { showBinaryChoiceDialog } from "../common/vscode/dialog";
 import { getOnDiskWorkspaceFolders } from "../common/vscode/workspace-folders";
 import { displayQuickQuery } from "./quick-query";
@@ -27,6 +26,7 @@ import { ResultsView } from "./results-view";
 import { DatabaseItem, DatabaseManager } from "../databases/local-databases";
 import {
   createInitialQueryInfo,
+  createTimestampFile,
   getQuickEvalContext,
   promptUserToSaveChanges,
   QueryOutputDir,
